FTC: upstream: Add the ability to use rsync as download protocol
This allows for packages to be developed locally in another directory
and then copied in before you commit or push to the official
repository.
The SUBDIR is _not_ removed so rsync has something to work with.
You will likely want to have your SCM (git or svn) to ignore the SUBDIR.

Signed-off-by: Jimi Xenidis <jimix@pobox.com>